+++ platform/plugins/documentation/database/migrations/2025_04_17_114008_create_articles_table.php
@@ -11,21 +11,37 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('articles', function (Blueprint $table) {
-            $table->id();
-            $table->string('title', 255);
-            $table->text('content');
-            $table->unsignedBigInteger('documentation_id');
-            $table->unsignedBigInteger('topic_id');
-            $table->unsignedBigInteger('user_id');//created_by
-            $table->string('status', 60)->default('published');
-            $table->timestamps();
+        if (! Schema::hasTable('articles')) {
+            Schema::create('articles', function (Blueprint $table) {
+                $table->id();
+                $table->string('title', 255);
+                $table->text('content');
+                $table->unsignedBigInteger('documentation_id');
+                $table->unsignedBigInteger('topic_id');
+                $table->unsignedBigInteger('user_id'); //created_by
+                $table->string('status', 60)->default('published');
+                $table->timestamps();
 
-            $table->foreign('documentation_id')->references('id')->on('documentations')->onDelete('cascade');
-            $table->foreign('topic_id')->references('id')->on('topics')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            
-        });
+                $table->foreign('documentation_id')->references('id')->on('documentations')->onDelete('cascade');
+                $table->foreign('topic_id')->references('id')->on('topics')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
+        if (! Schema::hasTable('articles_translations')) {
+            Schema::create('articles_translations', function (Blueprint $table) {
+                $table->string('lang_code');
+                $table->unsignedBigInteger('article_id');
+                $table->string('title', 255)->nullable();
+                $table->text('content')->nullable();
+
+                $table->primary(['lang_code', 'article_id'], 'articles_translations_primary');
+
+                $table->foreign('article_id')
+                    ->references('id')
+                    ->on('articles')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -34,5 +50,6 @@ return new class extends Migration
     public function down(): void
     {
         Schema::dropIfExists('articles');
+        Schema::dropIfExists('articles_translations');
     }
 };
